requirements.txt
api==0.0.7
appdirs==1.4.3
asgiref==3.2.3
astroid==2.4.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.11.0
filelock==3.0.12
fuzzy-c-means==0.0.6
geographiclib==1.50
geopy==1.22.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
nose==1.3.7
numpy==1.18.4
pandas==1.0.3
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.4
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-fuzzy==0.4.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.5
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.0.0
